compile fixes for linux/gcc4
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@39379 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ tools/switchboard2/main.cc
@@ -19,6 +19,8 @@
 
 #include	<sys/types.h>
 #include	<sys/socket.h>
+#include	<sys/un.h>
+#include	<sys/wait.h>
 
 #include	<netinet/in.h>
 #include	<arpa/inet.h>
@@ -350,7 +352,7 @@ dostat:
 void *
 acceptor_thread(void *arg)
 {
-	int fd = reinterpret_cast<int>(arg);
+	int fd = reinterpret_cast<intptr_t>(arg);
 	int newfd;
 	struct sockaddr addr;
 	socklen_t addrlen = sizeof(addr);
